# openapi_updated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Social Trends API",
    "description": "API per aggregare trend da TikTok e Instagram",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/auth/v2/register": {
      "post": {
        "tags": [
          "🔐 Authentication V2"
        ],
        "summary": "Register User",
        "description": "🔐 **Registrazione Utente con Verifica Email**\n\nRegistra un nuovo utente e invia email di verifica.\nSolo dopo la verifica email riceverai la tua API key.\n\n**Piani disponibili:**\n- **Free**: 1.000 chiamate/mese, solo endpoint /global\n- **Developer**: 10.000 chiamate/mese, richiede upgrade a pagamento\n- **Business**: 50.000 chiamate/mese, richiede upgrade a pagamento\n- **Enterprise**: 200.000 chiamate/mese, richiede upgrade a pagamento",
        "operationId": "register_user_v1_auth_v2_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRegistrationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/v2/verify-email": {
      "get": {
        "tags": [
          "🔐 Authentication V2"
        ],
        "summary": "Verify Email",
        "description": "✅ **Verifica Email**\n\nVerifica la tua email e ricevi la tua API key.\nQuesto endpoint viene chiamato automaticamente quando clicchi il link nell'email.",
        "operationId": "verify_email_v1_auth_v2_verify_email_get",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailVerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/v2/my-account": {
      "get": {
        "tags": [
          "🔐 Authentication V2"
        ],
        "summary": "Get Account Info",
        "description": "👤 **Informazioni Account**\n\nMostra tutte le informazioni del tuo account e delle tue API keys.",
        "operationId": "get_account_info_v1_auth_v2_my_account_get",
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "La tua API key",
              "title": "Api Key"
            },
            "description": "La tua API key"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Account Info V1 Auth V2 My Account Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailVerificationResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "EmailVerificationResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserRegistrationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "tier": {
            "type": "string",
            "pattern": "^(free|developer|business|enterprise)$",
            "title": "Tier",
            "default": "free"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "UserRegistrationRequest"
      },
      "UserRegistrationResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "requires_email_verification": {
            "type": "boolean",
            "title": "Requires Email Verification",
            "default": false
          },
          "verification_sent_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Sent To"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "UserRegistrationResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}